[refactor] move PageContainer to layouts category
