+++ Framework/UI/Layouts/Page.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import { useTheme, Appbar } from 'react-native-paper';
-import { LinearLayout } from '../Layouts/Layouts';
+import { LinearLayout } from './Layouts';
 import { onLocalDataUpdate } from '../../Contexts/LocalDataContext';
 import { padSize } from '../../Common/Values';
